requirements.txt
requests
python-dotenv
pandas
numpy
statsmodels
scikit-learn
tensorflow
streamlit
plotly
statsmodels>=0.14.0
streamlit-autorefresh
